+++ packages/lib/src/actionMiddlewares/onActionMiddleware.ts
@@ -91,10 +91,20 @@ function actionContextToActionCall(ctx: SimpleActionContext): ActionCall {
   }
 }
 
+const dateKey = "$dateAsTimestamp"
+const mapKey = "$mapAsArray"
+const setKey = "$setAsArray"
+
 /**
  * Transforms an action call argument by returning its serializable equivalent.
- * In more detail, this will return the snapshot of models, the non observable equivalent of observable values,
- * or if it is a primitive then the primitive itself.
+ * In more detail, this will transform:
+ * - Primitives as is.
+ * - Models / other tree nodes into their snapshots.
+ * - The non observable equivalent of observable values.
+ * - Dates as an object like `{ $dateAsTimestamp: number }`.
+ * - Maps as an object like `{ $mapAsArray: [[ ... ]] }`
+ * - Sets as an object like `{ $setAsArray: [ ... ] }`
+ *
  * If the value cannot be serialized it will throw an exception.
  *
  * @param argValue Argument value to be transformed into its serializable form.
@@ -104,16 +114,56 @@ export function serializeActionCallArgument(argValue: any): any {
   if (isPrimitive(argValue)) {
     return argValue
   }
+
   if (isTweakedObject(argValue, true)) {
     return getSnapshot(argValue)
   }
 
   const origValue = argValue
+
+  if (argValue instanceof Date) {
+    return { [dateKey]: +argValue }
+  }
+
   if (isObservable(argValue)) {
     argValue = toJS(argValue, { exportMapsAsObjects: false, detectCycles: false })
   }
-  if (isPlainObject(argValue) || Array.isArray(argValue)) {
-    return argValue
+
+  if (Array.isArray(argValue)) {
+    return argValue.map(serializeActionCallArgument)
+  }
+
+  if (argValue instanceof Map) {
+    const arr: [any, any][] = []
+
+    const iter = argValue.keys()
+    let cur = iter.next()
+    while (!cur.done) {
+      const k = cur.value
+      const v = argValue.get(k)
+      arr.push([serializeActionCallArgument(k), serializeActionCallArgument(v)])
+      cur = iter.next()
+    }
+
+    return { [mapKey]: arr }
+  }
+
+  if (argValue instanceof Set) {
+    const arr: any[] = []
+
+    const iter = argValue.keys()
+    let cur = iter.next()
+    while (!cur.done) {
+      const k = cur.value
+      arr.push(serializeActionCallArgument(k))
+      cur = iter.next()
+    }
+
+    return { [setKey]: arr }
+  }
+
+  if (isPlainObject(argValue)) {
+    return mapObjectFields(argValue, serializeActionCallArgument)
   }
 
   throw failure(`serializeActionCallArgument could not serialize the given value: ${origValue}`)
@@ -135,15 +185,62 @@ export function serializeActionCall(actionCall: ActionCall): ActionCall {
 
 /**
  * Transforms an action call argument by returning its deserialized equivalent.
- * In more detail, this will transform back the snapshot of models, and keep everything else as is.
- * If the value cannot be deserialized it will throw an exception.
+ * In more detail, this will transform:
+ * - The snapshot of models/tree nodes back into models/tree nodes.
+ * - `{ $dateAsTimestamp: number }` back to `Date` objects.
+ * - `{ $mapAsArray: [[ ... ]] }` back to `Map` objects.
+ * - `{ $setAsArray: [...] }` back to `Set` objects.
+ * - Everything else will be kept as is.
  *
  * @param argValue Argument value to be transformed into its deserialized form.
  * @returns The deserialized form of the passed value.
  */
 export function deserializeActionCallArgument(argValue: any): any {
+  if (isPrimitive(argValue)) {
+    return argValue
+  }
+
+  if (typeof argValue === "object" && typeof argValue[dateKey] === "number") {
+    return new Date(argValue[dateKey])
+  }
+
+  if (Array.isArray(argValue)) {
+    return argValue.map(deserializeActionCallArgument)
+  }
+
   if (isModelSnapshot(argValue)) {
     return fromSnapshot(argValue)
+  }
+
+  if (isPlainObject(argValue)) {
+    if (argValue[mapKey]) {
+      const arr: [any, any][] = argValue[mapKey]
+      const map = new Map()
+
+      const len = arr.length
+      for (let i = 0; i < len; i++) {
+        const k = arr[i][0]
+        const v = arr[i][1]
+        map.set(deserializeActionCallArgument(k), deserializeActionCallArgument(v))
+      }
+
+      return map
+    }
+
+    if (argValue[setKey]) {
+      const arr: any[] = argValue[setKey]
+      const set = new Set()
+
+      const len = arr.length
+      for (let i = 0; i < len; i++) {
+        const k = arr[i]
+        set.add(deserializeActionCallArgument(k))
+      }
+
+      return set
+    }
+
+    return mapObjectFields(argValue, deserializeActionCallArgument)
   }
 
   return argValue
@@ -161,4 +258,16 @@ export function deserializeActionCall(actionCall: ActionCall): ActionCall {
     ...actionCall,
     args: actionCall.args.map(deserializeActionCallArgument),
   }
+}
+
+function mapObjectFields(originalObj: any, mapFn: (x: any) => any): any {
+  const obj: any = {}
+  const keys = Object.keys(originalObj)
+  const len = keys.length
+  for (let i = 0; i < len; i++) {
+    const k = keys[i]
+    const v = originalObj[k]
+    obj[k] = mapFn(v)
+  }
+  return obj
 }